--- a/packages/gatsby/src/utils/cache-lmdb.ts
+++ b/packages/gatsby/src/utils/cache-lmdb.ts
@@ -1,4 +1,4 @@
-import { open, RootDatabase, Database } from "lmdb-store"
+import { open, RootDatabase, Database, DatabaseOptions } from "lmdb-store"
 import fs from "fs-extra"
 import path from "path"
 
@@ -16,12 +16,20 @@ const cacheDbFile =
 export default class GatsbyCacheLmdb {
   private static store
   private db: Database | undefined
+  private readonly encoding: DatabaseOptions["encoding"]
   public readonly name: string
   // Needed for plugins that want to write data to the cache directory
   public readonly directory: string
 
-  constructor({ name = `db` }: { name: string }) {
+  constructor({
+    name = `db`,
+    encoding = `msgpack`,
+  }: {
+    name: string
+    encoding: DatabaseOptions["encoding"]
+  }) {
     this.name = name
+    this.encoding = encoding
     this.directory = path.join(process.cwd(), `.cache/caches/${name}`)
   }
 
@@ -46,6 +54,7 @@ export default class GatsbyCacheLmdb {
     if (!this.db) {
       this.db = GatsbyCacheLmdb.getStore().openDB({
         name: this.name,
+        encoding: this.encoding,
       })
     }
     return this.db