--- a/packages/gatsby/src/utils/worker/__tests__/queries.ts
+++ b/packages/gatsby/src/utils/worker/__tests__/queries.ts
@@ -19,6 +19,7 @@ import { getDataStore } from "../../../datastore"
 import { IGroupedQueryIds } from "../../../services"
 import { IGatsbyPage } from "../../../redux/types"
 import { runQueriesInWorkersQueue } from "../pool"
+import { readPageQueryResult } from "../../page-data"
 
 let worker: GatsbyTestWorkerPool | undefined
 
@@ -252,8 +253,9 @@ describeWhenLMDB(`worker (queries)`, () => {
     if (!worker) fail(`worker not defined`)
     const stateFromWorker = await worker.single.getState()
 
-    const pageQueryResult = await fs.readJson(
-      `${stateFromWorker.program.directory}/.cache/json/_foo.json`
+    const pageQueryResult = await readPageQueryResult(
+      `${stateFromWorker.program.directory}/public`,
+      `/foo`
     )
 
     expect(pageQueryResult.data).toStrictEqual({
@@ -267,8 +269,9 @@ describeWhenLMDB(`worker (queries)`, () => {
     if (!worker) fail(`worker not defined`)
     const stateFromWorker = await worker.single.getState()
 
-    const pageQueryResult = await fs.readJson(
-      `${stateFromWorker.program.directory}/.cache/json/_bar.json`
+    const pageQueryResult = await readPageQueryResult(
+      `${stateFromWorker.program.directory}/public`,
+      `/bar`
     )
 
     expect(pageQueryResult.data).toStrictEqual({
@@ -289,8 +292,9 @@ describeWhenLMDB(`worker (queries)`, () => {
     const stateFromWorker = await worker.single.getState()
 
     // Called the complete ABC so we can test _a
-    const pageQueryResultA = await fs.readJson(
-      `${stateFromWorker.program.directory}/.cache/json/_a.json`
+    const pageQueryResultA = await readPageQueryResult(
+      `${stateFromWorker.program.directory}/public`,
+      `/a`
     )
 
     expect(pageQueryResultA.data).toStrictEqual({
@@ -299,8 +303,9 @@ describeWhenLMDB(`worker (queries)`, () => {
       },
     })
 
-    const pageQueryResultZ = await fs.readJson(
-      `${stateFromWorker.program.directory}/.cache/json/_z.json`
+    const pageQueryResultZ = await readPageQueryResult(
+      `${stateFromWorker.program.directory}/public`,
+      `/z`
     )
 
     expect(pageQueryResultZ.data).toStrictEqual({