## Description

This uses `lmdb-store` backed cache instance for storing/retrieving page query results (to generate final `page-data` file once everything is ready) instead of writing temporary files and reading them out later.

## Benchmarks

Before change (LMDB, PQR, 5 workers):

```
run queries in workers - 359.396s - 25848/25848 71.92/s
Writing page-data.json files to public directory - 77.187s - 25847/25847 334.86/s
Total for those activities: 436.583s
```

With this change (LMDB, PQR, 5 workers):

```
run queries in workers - 301.939s - 25848/25848 85.61/s
Writing page-data.json files to public directory - 79.732s - 25847/25847 324.17/s
Total for those activities: 381.671s
```

So for test site with 5 workers this yields ~12% speedup (for activities/steps that make use of page query results)

## Related Issues

[ch34445]