@simonjoom 2021-07-30
Why you not let LmdbStore take care about page-data by default if the compilation speed improve at 12%? 
personaly i removed the test datastore.isLmdbStore() to use LMDB_STORE only for page-data
 
I don't use the env GATSBY_EXPERIMENTAL_LMDB_STORE for my project