@simonjoom 2021-08-01
I found an issue in your code from this pr

`async function savePageQueryResult(programDir, pagePath, stringifiedResult) {
   if ((0, _datastore.isLmdbStore)()) {
    //isssue here no await
    //savePageQueryResultsPromise = getLMDBPageQueryResultsCache().set(pagePath, stringifiedResult);
    await getLMDBPageQueryResultsCache().set(pagePath, stringifiedResult);
  } else {
    const pageQueryResultsPath = _path.default.join(programDir, `.cache`, `json`, `${pagePath.replace(/\//g, `_`)}.json`);

    await _fsExtra.default.outputFile(pageQueryResultsPath, stringifiedResult);
  } 
}`